% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_bins.R
\docType{class}
\name{Bin-class}
\alias{Bin-class}
\title{Bin Abstract Class}
\description{
The Bin abstract class for all unit bin levels such as
\link{BinNumeric} \link{BinFactor}, \link{BinMissing}, and \link{BinException}.
}
