% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{combinable}
\alias{combinable}
\title{Test Whether Objects can be Combined}
\usage{
combinable(a, b)
}
\arguments{
\item{a}{First object.}

\item{b}{Second object.}
}
\description{
Test whether two objects are combinable.
}
\details{
Combinable takes two objects and returns TRUE or FALSE for whether
they can be combined. The \link{combine} function makes use of this test
to implement conditional logic.
}

