% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_bins.R, R/class_level.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,Bin,Bin-method}
\alias{combine,BinNumeric,BinNumeric-method}
\alias{combine,Level,Level-method}
\alias{combine,Level,list-method}
\alias{combine,List,Level-method}
\alias{combine,list,Level-method}
\title{Combine Objects}
\usage{
combine(a, b, ...)

\S4method{combine}{BinNumeric,BinNumeric}(a, b)

\S4method{combine}{Bin,Bin}(a, b)

\S4method{combine}{Level,Level}(a, b)

\S4method{combine}{list,Level}(a, b)

\S4method{combine}{Level,list}(a, b)
}
\arguments{
\item{a}{First object to be combined.}

\item{b}{Second object to be combined.}

\item{...}{Additional arguments passed on to other methods.}
}
\description{
Combine combinable objects such as \link{Bin-class} or \link{Level-class}
}
\details{
Combine takes two objects of the same class and combines them 
into one object of the same class.
}

