% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_bins.R
\docType{methods}
\name{expand}
\alias{expand}
\alias{expand,BinNumeric-method}
\alias{combineexpand,BinNumeric,BinNumeric-method}
\title{Expand an Object}
\usage{
expand(object, ...)

\S4method{expand}{BinNumeric}(object, breaks, ...)
}
\arguments{
\item{object}{An object that can be expanded.}

\item{...}{Additional arguments passed on to other methods.}

\item{breaks}{A numeric vector of two or more cut points.}
}
\description{
Expand an expandable object such as \link{Bin-class} or \link{Level-class}
}
\details{
Expand will expand an objecet if it can, returning multiple objects
of the same class. Otherwise it will return the input object.
}
