% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_bins.R
\docType{methods}
\name{get_boolean_mask}
\alias{get_boolean_mask}
\alias{get_boolean_mask,BinNumeric,numeric-method}
\alias{get_boolean_mask,BinNumeric,factor-method}
\alias{get_boolean_mask,BinFactor,factor-method}
\alias{get_boolean_mask,BinException,numeric-method}
\alias{get_boolean_mask,BinMissing,ANY-method}
\title{Get Boolean Mask}
\usage{
get_boolean_mask(object, x, ...)

\S4method{get_boolean_mask}{BinNumeric,numeric}(object, x, ..., mask = FALSE)

\S4method{get_boolean_mask}{BinFactor,factor}(object, x, ...)

\S4method{get_boolean_mask}{BinException,numeric}(object, x, ...)

\S4method{get_boolean_mask}{BinMissing,ANY}(object, x, ...)
}
\arguments{
\item{object}{An object}

\item{x}{A vector of the appropriate type to map to TRUE/FALSE.}

\item{...}{Additional arguments passed on to other methods.}
}
\description{
Returns a logical vector the same length as argument \code{x} indicating
membership in an appropriate interval or a match to an appropriate level.
}
