% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{info}
\alias{info}
\title{Get Information About an Object}
\usage{
info(object, ...)
}
\arguments{
\item{object}{Any \link{onx} object.}

\item{...}{Additional arguments passed on to other methods.}
}
\description{
Print descriptive information about an object and its slots/fields
}
