% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_level.R
\docType{class}
\name{Level-class}
\alias{Level-class}
\alias{LevelContinuous}
\alias{LevelContinuous-class}
\title{Level Class}
\description{
The level class consists of two main types: \code{LevelContinuous} and
\code{LeveLDiscrete}. Both differ in their operations and what bins subtypes they can
manage..

\code{LevelContinuous} object contain a list of bins each of which
belongs to the LevelContinuousAppendable class union. These bin types consist of
\code{BinNumeric}, \code{BinException}, and \code{BinMissing.}
}
\section{Slots}{

\describe{
\item{\code{bins}}{A list of bins contained in this level.}

\item{\code{values}}{A named list of values associated with the level.}
}}
\section{LevelContinuous}{

}

