% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{Sort}
\alias{Sort}
\title{Sort Objects by their OrderValue}
\usage{
Sort(object, ...)
}
\arguments{
\item{object}{First object.}

\item{...}{Additional arguments passed on to other methods.}
}
\description{
Sort a collection of objects by their \link{ordervalue}s
}

