% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_bins.R
\docType{methods}
\name{get_label}
\alias{get_label}
\alias{get_label,BinNumeric-method}
\title{Get Object Labels}
\usage{
get_label(object, ...)

\S4method{get_label}{BinNumeric}(object, digits = 3, ...)
}
\arguments{
\item{object}{An object that can generate labels.}

\item{...}{Additional arguments passed on to other methods.}

\item{digits}{rounding digits for numeric bin boundaries}
}
\description{
Some additional info
}
